refactor: move the per-service scheduler into its own module
The new scheduler pushed internal/engine/lifecycle/mod.rs to 510 code
lines, over the 500-line hard limit, and line-limit.yml failed the build.

Split by subject rather than by count: the scheduler is where the
difference between "levels" and "a dependency graph" lives, and the two
deadlock hazards it avoids — polling every future rather than the first
N, and never holding a permit across a dependency wait — are documented
inline where someone changing it will read them. Folding that into
run_up was always the wrong home for it.

468 and 79 lines. No behaviour change; ordering re-verified against real
Podman (a dependent still starts after its dependency).

Part of #1071
